+++ test/validation/variables_test.rb
@@ -15,6 +15,12 @@ class VariablesTest < Test::Unit::TestCase
       assert_equal true, (VALID_VARIABLE_TYPES.include?(JSON.parse(File.read(file))["type"]) rescue false)
     end
 
+    if (not [nil, ''].include?(JSON.parse(File.read(file))["domain"]) rescue false)
+      define_method("test_domain_exists: "+file) do
+        assert_equal true, (File.exists?(File.join("domains", JSON.parse(File.read(file))["domain"]+".json")) rescue false)
+      end
+    end
+
   end
 
 end