+++ CHANGELOG.md
@@ -38,4 +38,5 @@ Data Dictionary currently comprised of two primary folders: `variables` and `dom
 - Added configuration for TravisCI Continuous Integration tests
 - Added tests for JSON validity of variables and domains
 - Added test to check presence/validity of variable type
+- Added test to check if a domain referenced from a variable exists
 - Added a rake task to create CSVs of the JSON data dictionary